refactor(core): extract DivergenceTracker into own module
DivergenceTracker is a self-contained rate-threshold alarm
(state_root-mismatch rejection counter + cooldown logic). It lives
in `blockchain.rs` for historical reasons but has zero dependency
on the Blockchain struct or any chain state. The struct + impl +
tests together are ~120 lines that don't belong in the middle of
the chain orchestration file.

Pulled into `crate::divergence` (visibility `pub(crate)` matches
the original); blockchain.rs now imports it via `use
crate::divergence::DivergenceTracker;` and the field reference
stays unchanged. No external crate touches DivergenceTracker so
no external caller change needed.

The two tests (`test_divergence_tracker_counts`,
`test_divergence_alarm_cooldown`) moved with the type into
`divergence::tests`. Const visibility on `WINDOW_SECS`,
`ALARM_THRESHOLD`, `ALARM_COOLDOWN_SECS` widened to `pub(crate)`
so the moved tests can reference them via `Self::` — was implicit
through being in the same module before.

Same shape as the fork-heights extract (PR #634): refactor-only,
no behaviour change, no consensus path touched.

sentrix-core test suite: 213 passed (same count, 2 tests just
moved their home module).
